additional test; move cli tests to the docker subpackage
